Creating a robot factory.
Creating the interface, concrete, and return types for the robot factory. The interface, IRobotFactory, specifies the required methods for a robot factory. The concrete class, RobotFactory, provides a simple implementation following the requirements of the interface. The two type classes, Enemy and Teammate, extend the AbstractRobot class and provide two different representations for a robot.