Preparing for management classes.
Created an event and message handler interface for management classes. These interfaces require certain methods that help with the management of events and messages. By implementing these methods, the specifics of management classes could be ignored and all events and methods can be passed to event handlers and message handlers respectively.