Adding the robot manager class.
Created the RobotManager class that allows for easy management of enemy and teammate robots. This class provides basic functionality that can be used to retrieve robots by name and get a collection of all robots. The robot manager handles scanned robot and robot death events as well as the new class RobotSnapshotMessage. This new message allows teammates to send robot snapshots to each other.