Adding turnToLine methods. (#2)
Adding the ability for a robot to turn to a line. This allows the robot to turn so it is in line with a specified line. It does not matter which way the robot is headed on the line as it can move backwards just as face. The methods will return which heading was used to turn to the line.

Also added a few information methods pertaining to velocity and turn rate.